requirements.txt
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2.17
django-agnocomplete==1.0.0
django-appconf==1.0.5
django-compressor==2.4
django-environ==0.8.1
django-jsonfield==1.4.1
gunicorn==20.1.0
idna==2.10
oauthlib==3.2.0
Pillow==8.0.1
psycopg2==2.8.4
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.0.6
requests==2.23.0
requests-oauthlib==1.3.1
rjsmin==1.1.0
six==1.16.0
social-auth-app-django==3.1.0
social-auth-core==3.3.2
sqlparse==0.4.2
umap-project==1.2.3
urllib3==1.25.11
